feat(screen): bind endereco_input to on_locate_text event (master)
Bind the endereco_input field to the on_locate_text event in the screen.py file. This allows for the description of the location to be retrieved when the location code is entered. If an exception is raised, it is printed and ignored. The commit is made on the master branch.